+++ lib/td_ie/canada/abilities.ex
@@ -4,6 +4,8 @@ defmodule TdIe.Canada.Abilities do
   alias TdIe.Canada.IngestAbilities
   alias TdIe.Ingests.Ingest
   alias TdIe.Ingests.IngestVersion
+  alias TdIe.Canada.LinkAbilities
+  alias TdCache.Link
 
   defimpl Canada.Can, for: User do
     # administrator is superpowerful
@@ -81,6 +83,18 @@ defmodule TdIe.Canada.Abilities do
           %IngestVersion{} = ingest_version
         ) do
       IngestAbilities.can?(user, :view_ingest, ingest_version)
+    end
+
+    def can?(%User{} = user, action, %Link{} = link) do
+      LinkAbilities.can?(user, action, link)
+    end
+
+    def can?(%User{} = user, :create_link, %{ingest: ingest}) do
+      LinkAbilities.can?(user, :create_link, ingest)
+    end
+
+    def can?(%User{} = user, action, %{hint: :link} = resource) do
+      LinkAbilities.can?(user, action, resource)
     end
 
     def can?(%User{is_admin: true}, _action, %{}) do